back_end/requirements.txt
flask
flask-restful
pytesseract
openai
firebase-admin
python-dotenv
Pillow 
requests 
PyPDF2 
python-docx 